Restore SolverBasedTest0 to JUnit4 so as not to break the API
Internally, we'll use a new class `SolverBasedTest` (without the `0` at the end) for testing with JUnit6. The new class won't be included in the *.jar and is not part of the official API for now
