+++ src/org/sosy_lab/java_smt/test/FloatingPointFormulaManagerTest.java
@@ -52,7 +52,7 @@ import org.sosy_lab.java_smt.api.SolverException;
 import org.sosy_lab.java_smt.api.visitors.FormulaVisitor;
 
 public class FloatingPointFormulaManagerTest
-    extends SolverBasedTest0.ParameterizedSolverBasedTest0 {
+    extends SolverBasedTest.ParameterizedSolverBasedTest {
 
   // numbers are small enough to be precise with single precision
   private static final int[] SINGLE_PREC_INTS = new int[] {0, 1, 2, 5, 10, 20, 50, 100, 200, 500};
