+++ src/org/sosy_lab/java_smt/test/ModelEvaluationTest.java
@@ -29,7 +29,7 @@ import org.sosy_lab.java_smt.api.SolverException;
 import org.sosy_lab.java_smt.basicimpl.AbstractStringFormulaManager;
 
 /** Test that we can request evaluations from models. */
-public class ModelEvaluationTest extends SolverBasedTest0.ParameterizedSolverBasedTest0 {
+public class ModelEvaluationTest extends SolverBasedTest.ParameterizedSolverBasedTest {
 
   /**
    * This is the default boolean value for unknown model evaluations. For unknown model evaluation
