+++ src/org/sosy_lab/java_smt/test/UserPropagatorTest.java
@@ -21,7 +21,7 @@ import org.sosy_lab.java_smt.api.SolverException;
 import org.sosy_lab.java_smt.api.UserPropagator;
 import org.sosy_lab.java_smt.basicimpl.AbstractUserPropagator;
 
-public class UserPropagatorTest extends SolverBasedTest0.ParameterizedSolverBasedTest0 {
+public class UserPropagatorTest extends SolverBasedTest.ParameterizedSolverBasedTest {
 
   @BeforeEach
   public void init() {
